Clean out old Marshall class (rename to Helper) and rename FieldMarshall.cs.
