+++ H5_DataPipeline/H5_DataPipeline.csproj
@@ -98,10 +98,10 @@
   <ItemGroup>
     <Compile Include="Assistants\Clanalyzer.cs" />
     <Compile Include="Assistants\Historian.cs" />
-    <Compile Include="Assistants\FieldMarshall.cs" />
+    <Compile Include="Assistants\Marshall.cs" />
     <Compile Include="Assistants\HistorianScribe.cs" />
     <Compile Include="Assistants\Mortician.cs" />
-    <Compile Include="DeleteTheseSoon\Marshall.cs" />
+    <Compile Include="Assistants\Helper.cs" />
     <Compile Include="Models\t_h5matchesExtensions.cs" />
     <Compile Include="Models\t_h5matches_playersformatchExtensions.cs" />
     <Compile Include="Models\t_h5matches_ranksandscoresExtension.cs" />
